requirements.txt
openai
python-dotenv
Pillow
requests

